backend/pyproject.toml
[project]
name = "shipping-labels"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.4",
    "django>=5.1",
    "django-cors-headers>=4.4",
    "django-environ>=0.11",
    "djangorestframework>=3.15",
    "drf-spectacular>=0.27",
    "httpx>=0.27",
    "psycopg[binary]>=3.2",
    "redis>=5.0",
    "structlog>=24.0",
]

[tool.uv]
dev-dependencies = [
    "factory-boy",
    "freezegun",
    "pytest",
    "pytest-django",
    "ruff",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
[tool.setuptools.packages.find]
where = ["."]
include = ["config", "core", "imports", "shipments", "addresses", "shipping"]

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py310"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "W191", # indentation contains tabs
    "B904", # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.pytest.ini_options]
log_cli = true
asyncio_mode = "auto"
